Load stages with issues from multiple repos
Issues are fetched on a per-repo basis to minimize the number of requests, and
each call only fetches issues that are tagged with any of the flow labels. For
now, flow labels are assumed to come from only the first repo in your list, but
something like a union would be straightforward. The `GitHUD.Kanban` module is
responsible for managing label/issue fetching and rendering.

Because multiple repos are now present, issue slugs take the form
`issue-repo-number` to disambiguate between two repos that have an issue with
the same number.

To distinguish between the issues attached to a `Repo` and those attached to a
`Label`, there are now two types of issue list: `RepoIssues` and `LabelIssues`.
I'm not sure if there's a better way to go about that. `RepoIssues` expect to
receive a `repo` and use it to generate a URL for `fetch`ing. `LabelIssues` are
intended only as a store and cannot be `fetch`ed.

The resizing logic was moved into `GitHUD.Util` for easy shared access.