Prevent stage headers from varying in height
The UI is still reasonable on a 15" widescreen laptop with 7 columns, but
they're getting a little tight. Perhaps a minimum width is in order, wich
horizontal scrolling on the whole content. Or perhaps 7 columns is an
unreasonably long flow.

In case you miss the full title, it's available on mouseover.