Allow adding new repos without having to refresh
Now when you add a repo to the view, its issues will immediately appear. If it's
the first repo, its flow labels will be pulled down, and every subsequent issue
will only use those labels for now.

Backbone uses `Model`s' ID attribute to determine whether it's already present
in a collection, and `Issue` IDs were colliding. So now the `GitHUD.Issue` ID is
a full slug instead of just its number.

The decision on whether a new repo should be added is now in Core, and we're now
more consistent about listening for `add` events in anything that cares.