Enable drag + drop between stages
jQuery UI handles the dragging and sorting, and all the callback has to do is
remove the relevant model/view from its list. I'd've liked to use a
`Backbone.Collection` for `stageViews`, but it complained with a cryptic message.

Slugs got a makeover to be more readable/parseable since HTML appears happy with
whatever characters you put in there except for whitespace.

Next: Notify GitHub