Write stage changes back to GitHub
I fought with Backbone and GitHub's not-quite-RESTful `issue/labels` endpoint
for quite a while, but ultimately direct jQuery was the easier way to go. It's a
shame to have nonuniform code talking to their API, but using Backbone was long,
hackey and buggy, so I'll settle for short, unconventional and functional.