Update to GitHub's new /issues behavior
Last night, the `/issues` endpoint began returning issues that have *all* of the
labels in the `labels` parameter, instead of *any*. This was apparently the
intended behavior all along (to parallel the Issues UI), and the docs leave it
ambiguous that the old behavior seemed reasonable when I found it.

It does simplify the code somewhat, though it results in `Labels * Repos` API
calls instead of just `Repos`.